move collections_local_copy to fle_utils
