+++ kalite/basetests/tests.py
@@ -227,7 +227,6 @@ class PackagesTests(DependenciesTests):
         "smmap": "0.8.2",
         "tastypie": "0.11.0",
         "youtube_dl": "2014.12.10.3",
-        "collections_local_copy": NO_VERSION,
         "memory_profiler": "0.26",
         "mimeparse": "0.1.4",
         "oauth": "1.0",
