+++ kalite/i18n/base.py
@@ -3,7 +3,7 @@ import os
 import re
 import requests
 import shutil
-from collections_local_copy import OrderedDict
+from fle_utils.collections_local_copy import OrderedDict
 from fle_utils.internet.webcache import invalidate_web_cache
 
 from django.http import HttpRequest
